+++ CHANGELOG.md
@@ -88,3 +88,4 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Fixed function breakpoint logic, will not double break in case some source breakpoint is also set to first line of method.
 - Fixed undefined behavior in evaluation code.
 - Fixed memory leaks in process creation and generic evaluation code.
+- Fixed partial path matches.